Add NuGet cache
We want to be able to run the build inside the Docker container as the
same user that is running the scripts in VSTS. This will prevent files
being owned by root.

We'll map a folder from the host inside the container so that the build
can leverage the same NuGet cache even if we change the build container.